@@ -2,43 +2,46 @@
   "name": "javascript-development-environment",
   "version": "1.0.0",
   "description": "JavaScript development environment Pluralsight course by Cory House",
-  "scripts": {},
+  "scripts": {
+    "prestart": "babel-node sourceScripts/startMessage.js",
+    "start": "npm-run-all  webpack-dev source-start" ,
+    "source-start": "babel-node  sourceScripts/server.js",
+    "security-check": "npm audit",
+    "share": "npm-run-all --parallel start source-share",
+    "source-share": "lt -8080"
+  },
   "author": "Cory House",
   "license": "MIT",
   "dependencies": {
-    "whatwg-fetch": "1.0.0"
+    "whatwg-fetch": "3.0.0"
   },
   "devDependencies": {
-    "babel-cli": "6.16.0",
-    "babel-core": "6.17.0",
-    "babel-loader": "6.2.5",
-    "babel-preset-latest": "6.16.0",
-    "babel-register": "6.16.3",
-    "chai": "3.5.0",
-    "chalk": "1.1.3",
-    "cheerio": "0.22.0",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.2",
+    "babel-preset-env": "^1.7.0",
+    "chai": "4.2.0",
+    "chalk": "2.4.1",
+    "cheerio": "1.0.0-rc.2",
     "compression": "^1.7.3",
-    "cross-env": "3.1.3",
-    "css-loader": "0.25.0",
-    "eslint": "3.8.1",
+    "cross-env": "5.2.0",
+    "css-loader": "1.0.1",
+    "eslint": "5.9.0",
     "eslint-plugin-import": "^2.14.0",
-    "eslint-watch": "2.1.14",
+    "eslint-watch": "4.0.2",
     "express": "^4.16.4",
-    "extract-text-webpack-plugin": "1.0.1",
-    "html-webpack-plugin": "2.22.0",
-    "jsdom": "9.8.0",
-    "json-schema-faker": "^0.4.7",
+    "extract-text-webpack-plugin": "3.0.2",
+    "html-webpack-plugin": "3.2.0",
+    "jsdom": "13.0.0",
+    "json-schema-faker": "^0.5.0-rc16",
     "json-server": "^0.14.0",
     "localtunnel": "^1.9.1",
     "mocha": "^5.2.0",
     "nock": "^10.0.2",
-    "npm-run-all": "3.1.1",
-    "numeral": "1.5.3",
-    "rimraf": "2.5.4",
-    "style-loader": "0.13.1",
-    "webpack": "1.13.2",
-    "webpack-dev-middleware": "1.8.4",
-    "webpack-hot-middleware": "2.13.0",
-    "webpack-md5-hash": "0.0.5"
+    "npm-run-all": "4.1.5",
+    "numeral": "2.0.6",
+    "rimraf": "2.6.2",
+    "style-loader": "0.23.1",
+    "webpack": "^3.8.1",
+    "webpack-dev-server": "^2.9.7"
   }
 }
